+++ libft/src/ft_itoa.c
@@ -6,58 +6,50 @@
 /*   By: pibouill <pibouill@student.42prague.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/19 12:37:23 by pibouill          #+#    #+#             */
-/*   Updated: 2024/04/21 14:19:41 by pibouill         ###   ########.fr       */
+/*   Updated: 2024/12/18 12:55:28 by pibouill         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "../inc/libft.h"
 
 /*
 ** Returns the (char *) representation of int n
 */
 
-static int	nb_len(int n)
+static int	ft_len(int n)
 {
-	int	i;
+	int	l;
 
-	i = 0;
-	if (n == 0)
-		return (1);
+	l = 1;
 	if (n < 0)
-		i++;
-	while (n != 0)
 	{
-		n /= 10;
-		i++;
+		n *= -1;
+		l++;
 	}
-	return (i);
+	while (n / 10 > 0)
+	{
+		l++;
+		n = n / 10;
+	}
+	return (l);
 }
 
 char	*ft_itoa(int n)
 {
-	char		*res;
-	int			len;
-	long		tmp;
+	char	*res;
+	int		l;
 
-	len = nb_len(n);
-	tmp = n;
-	if (n < 0)
-		tmp *= -1;
-	res = malloc(sizeof(char) * (len + 1));
-	if (res == NULL)
+	l = ft_len(n);
+	res = (char *)malloc((l + 1) * sizeof(char));
+	if (!res)
 		return (NULL);
-	res[len] = 0;
-	if (tmp == 0)
-		res[0] = '0';
-	else
+	res[l] = '\0';
+	l--;
+	while (l >= 0)
 	{
-		while (len-- && tmp != 0)
-		{
-			res[len] = (tmp % 10) + '0';
-			tmp /= 10;
-		}
-		if (n < 0)
-			res[len] = '-';
+		res[l] = (n % 10) + '0';
+		n = n / 10;
+		l--;
 	}
 	return (res);
 }
