+++ libft/src/ft_strncmp.c
@@ -10,27 +10,29 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "../inc/libft.h"
 
 /*
 ** Returns first difference between *s1 and *s2
 ** Within n bytes
 */
 
-int	ft_strncmp(const char *s1, const char *s2, size_t n)
+int	ft_strncmp(char *str1, char *str2, size_t n)
 {
-	size_t	i;
+	unsigned char	*s1;
+	unsigned char	*s2;
 
-	i = 0;
-	if (n == 0)
-		return (0);
-	while (s1[i] && i < n - 1)
+	s1 = (unsigned char *)str1;
+	s2 = (unsigned char *)str2;
+	while ((*s1 || *s2) && *s1 == *s2 && n - 1 > 0)
 	{
-		if (s1[i] != s2[i])
-			return ((unsigned char)s1[i] - (unsigned char)s2[i]);
-		i++;
+		n--;
+		s1++;
+		s2++;
 	}
-	return ((unsigned char)s1[i] - (unsigned char)s2[i]);
+	if (n != 0)
+		return (*s1 - *s2);
+	return (0);
 }
 
 // int	main()
